part2/02-streamlit/special-mission/requirements.txt
addict==2.4.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
brotlipy==0.7.0
cachetools==5.1.0
certifi==2021.5.30
cffi @ file:///home/builder/ci_310/cffi_1642753365720/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==7.1.2
cryptography @ file:///home/builder/ci_310/cryptography_1640808721005/work
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mmcls==0.23.0
mmcv-full==1.5.1
-e git+https://github.com/seungriyou/Boostcamp-AI-Tech-Product-Serving.git@7d1025efd2c50ad95c23023c3653be98a53e3422#egg=mmsegmentation&subdirectory=part2/02-streamlit/special-mission/mmsegmentation
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.7.1
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.1.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.2.1.post0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.1
tzlocal==4.2
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650637206367/work
validators==0.19.0
watchdog==2.1.8
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
widgetsnbextension==3.6.0
yapf==0.32.0
